Add files via upload
In this class, focus is on the concept of physics engine, and how we use it to create objects that are close to real world objects. We observe here that just creating a physics object is not enough, we need to render those objects using various shape commands on the canvas. We can also customize different properties like restitution, density, friction, isStatic, etc. rectMode() and ellipseMode() are saviours saving us from extra calculations.